requirements.txt
fastapi
uvicorn
python-dotenv
pydantic>=2.0
psycopg2-binary
requests
sqlglot
openai
httpx
pytest
